+++ extensions/slack/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Slack Changelog
 
+## [New command to copy user email address] - {PR_MERGE_DATE}
+
+- Added a command to copy a user's email address through Slack. You may need to reauth to Slack for this to work.
+
 ## [Fix YAML manifest errors in README] - 2025-06-25
 
 ## [Fix search messages from specific user] - 2025-06-23